+++ extension/js/common/composer/composer-pwd-or-pubkey-container.ts
@@ -6,16 +6,18 @@ import { ComposerComponent } from './interfaces/composer-component.js';
 import { Ui } from '../browser.js';
 import { RecipientStatuses, SendBtnTexts } from './interfaces/composer-types.js';
 import { KeyImportUi } from '../ui/key_import_ui.js';
+import { Store } from '../platform/store.js';
 
 export class ComposerPwdOrPubkeyContainer extends ComposerComponent {
 
   private keyImportUI = new KeyImportUi({});
 
-  public initActions() {
+  public async initActions() {
     this.composer.S.cached('input_password').keyup(Ui.event.prevent('spree', () => this.showHideContainerAndColorSendBtn()));
     this.composer.S.cached('input_password').focus(() => this.showHideContainerAndColorSendBtn());
     this.composer.S.cached('input_password').blur(() => this.showHideContainerAndColorSendBtn());
-    if (this.composer.app.storageGetHideMsgPassword()) {
+    const store = await Store.getAcct(this.urlParams.acctEmail, ['hide_message_password']);
+    if (store.hide_message_password) {
       this.composer.S.cached('input_password').attr('type', 'password');
     }
   }
