+++ extension/js/common/composer/interfaces/composer-app-functions.ts
@@ -2,29 +2,16 @@
 
 'use strict';
 
-import { Subscription, ContactUpdate, DbContactObjArg, SendAsAlias } from '../../platform/store.js';
-import { KeyInfo, Contact } from '../../core/pgp.js';
-import { ProviderContactsQuery, SendableMsg } from '../../api/email_provider_api.js';
+import { SendAsAlias } from '../../platform/store.js';
+import { Contact } from '../../core/pgp.js';
+import { SendableMsg } from '../../api/email_provider_api.js';
 import { GmailRes } from '../../api/google.js';
 import { ProgressCb, ChunkedCb } from '../../api/api.js';
 import { Att } from '../../core/att.js';
-import { PubkeyResult } from './composer-types.js';
 import { Dict } from '../../core/common.js';
 
 export interface ComposerAppFunctionsInterface {
   doesRecipientHaveMyPubkey: (email: string) => Promise<boolean | undefined>;
-  storageGetAddresses: () => Dict<SendAsAlias> | undefined;
-  storageGetHideMsgPassword: () => boolean;
-  storageGetSubscription: () => Promise<Subscription>;
-  storageGetKey: (acctEmail: string, senderEmail: string) => Promise<KeyInfo>;
-  storageSetDraftMeta: (storeIfTrue: boolean, draftId: string, threadId: string, recipients?: string[], subject?: string) => Promise<void>;
-  storagePassphraseGet: (senderKi?: KeyInfo) => Promise<string | undefined>;
-  storageAddAdminCodes: (shortId: string, codes: string[]) => Promise<void>;
-  storageContactGet: (email: string[]) => Promise<(Contact | undefined)[]>;
-  storageContactUpdate: (email: string | string[], update: ContactUpdate) => Promise<void>;
-  storageContactSave: (contact: Contact) => Promise<void>;
-  storageContactSearch: (query: ProviderContactsQuery) => Promise<Contact[]>;
-  storageContactObj: (o: DbContactObjArg) => Promise<Contact>;
   emailProviderDraftGet: (draftId: string) => Promise<GmailRes.GmailDraftGet | undefined>;
   emailProviderDraftCreate: (acctEmail: string, mimeMsg: string, threadId?: string) => Promise<GmailRes.GmailDraftCreate>;
   emailProviderDraftUpdate: (draftId: string, mimeMsg: string) => Promise<GmailRes.GmailDraftUpdate>;
@@ -37,8 +24,5 @@ export interface ComposerAppFunctionsInterface {
   renderHelpDialog: () => void;
   factoryAtt: (att: Att, isEncrypted: boolean) => string;
   closeMsg: () => void;
-  whenMasterPassphraseEntered: (secondsTimeout?: number) => Promise<string | undefined>;
-  lookupPubkeyFromDbOrKeyserverAndUpdateDbIfneeded: (email: string) => Promise<Contact | "fail">;
-  collectAllAvailablePublicKeys: (senderEmail: string, senderKi: KeyInfo, recipients: string[]) => Promise<{ armoredPubkeys: PubkeyResult[], emailsWithoutPubkeys: string[] }>;
   updateSendAs: (sendAs: Dict<SendAsAlias>) => void;
 }
